Move difficulty package from bmath to beatmap
