+++ app/rulesets/osu/circle.go
@@ -1,8 +1,8 @@
 package osu
 
 import (
+	"github.com/wieku/danser-go/app/beatmap/difficulty"
 	"github.com/wieku/danser-go/app/beatmap/objects"
-	"github.com/wieku/danser-go/app/bmath/difficulty"
 	"math"
 )
